bug fix benchmarking
