+++ benchmarks/transport_montecarlo_packet.py
@@ -27,15 +27,6 @@ class BenchmarkMontecarloMontecarloNumbaPacket(BenchmarkBase):
     """
 
     @property
-    def geometry(self):
-        return NumbaRadial1DGeometry(
-            r_inner=np.array([6.912e14, 8.64e14], dtype=np.float64),
-            r_outer=np.array([8.64e14, 1.0368e15], dtype=np.float64),
-            v_inner=np.array([-1, -1], dtype=np.float64),
-            v_outer=np.array([-1, -1], dtype=np.float64),
-        )
-
-    @property
     def model(self):
         return numba_interface.NumbaModel(
             time_explosion=5.2e7,
@@ -57,74 +48,6 @@ class BenchmarkMontecarloMontecarloNumbaPacket(BenchmarkBase):
             bf_heating_estimator=np.empty((0, 0), dtype=np.float64),
             stim_recomb_cooling_estimator=np.empty((0, 0), dtype=np.float64),
             photo_ion_estimator_statistics=np.empty((0, 0), dtype=np.int64),
-        )
-
-    @parameterize(
-        {
-            "Packet params": [
-                {"mu": 0.3, "r": 7.5e14},
-                {"mu": -0.3, "r": 7.5e13},
-                {"mu": -0.3, "r": 7.5e14},
-            ]
-        }
-    )
-    def time_calculate_distance_boundary(self, packet_params):
-        mu = packet_params["mu"]
-        r = packet_params["r"]
-
-        calculate_distances.calculate_distance_boundary(
-            r, mu, self.geometry.r_inner[0], self.geometry.r_outer[0]
-        )
-
-    @parameterize(
-        {
-            "Parameters": [
-                {
-                    "packet": {"nu_line": 0.1, "is_last_line": True},
-                    "expected": None,
-                    "enable_full_relativity": True,
-                },
-                {
-                    "packet": {"nu_line": 0.2, "is_last_line": False},
-                    "expected": None,
-                    "enable_full_relativity": True,
-                },
-                {
-                    "packet": {"nu_line": 0.5, "is_last_line": False},
-                    "expected": utils.MonteCarloException,
-                    "enable_full_relativity": False,
-                },
-                {
-                    "packet": {"nu_line": 0.6, "is_last_line": False},
-                    "expected": utils.MonteCarloException,
-                    "enable_full_relativity": False,
-                },
-            ]
-        }
-    )
-    def time_calculate_distance_line(self, parameters):
-        packet_params = parameters["packet"]
-        nu_line = packet_params["nu_line"]
-        is_last_line = packet_params["is_last_line"]
-        enable_full_relativity = parameters["enable_full_relativity"]
-
-        time_explosion = self.model.time_explosion
-
-        doppler_factor = frame_transformations.get_doppler_factor(
-            self.static_packet.r,
-            self.static_packet.mu,
-            time_explosion,
-            enable_full_relativity
-        )
-        comov_nu = self.static_packet.nu * doppler_factor
-
-        calculate_distances.calculate_distance_line(
-            self.static_packet,
-            comov_nu,
-            is_last_line,
-            nu_line,
-            time_explosion,
-            enable_full_relativity
         )
 
     @parameterize(
